requirements.txt
cryptography==2.8
requests==2.22.0
discord.py==1.6.0
beautifulsoup4==4.9.3
python-dotenv==0.15.0
flask==1.1.2
matplotlib==3.4.2